Added AnnotationChart composition.
