+++ ts/Extensions/Annotations/AnnotationSeries.d.ts
@@ -4,9 +4,10 @@
  *
  * */
 
-import type { AnnotationSeries } from './Annotation';
+import type AnnotationChart from './AnnotationChart';
 import type MockPoint from './MockPoint';
 import type Point from '../../Core/Series/Point';
+import type Series from '../../Core/Series/Series';
 
 /* *
  *
@@ -19,6 +20,11 @@ export interface AnnotationPoint extends Point {
 }
 
 export type AnnotationPointType = (MockPoint|AnnotationPoint);
+
+export interface AnnotationSeries extends Series {
+    chart: AnnotationChart;
+    points: Array<AnnotationPoint>;
+}
 
 /* *
  *
