+++ ts/Extensions/Annotations/Popup/PopupAnnotations.ts
@@ -18,7 +18,7 @@
  *
  * */
 
-import type { AnnotationChart } from '../Annotation';
+import type AnnotationChart from '../AnnotationChart';
 import type AnnotationOptions from '../AnnotationOptions';
 import type { HTMLDOMElement } from '../../../Core/Renderer/DOMElementType';
 import type Popup from './Popup';
